Revert "upadated data fetching with LINQ"
This reverts commit dcd4631b0c7aaad34e965f1fb72da8654ec254ba.

+++ ExordiumApp/Assets/Scripts/SerializableClasses.cs
@@ -10,7 +10,7 @@ public class Item
     public string name;
     public string description;
     public string image_url;
-    public float price;
+    public decimal price;
     public int retailer_id;
     public int item_category_id;
 }
